requirements.txt
appdirs==1.4.3
beautifulsoup4==4.6.0
billboard.py==4.2.0
certifi==2018.1.18
chardet==3.0.4
click==6.7
colorama==0.2.5
command-not-found==0.3
Fabric3==1.13.1.post1
flake8==3.5.0
Flask==0.12.2
Flask-Cors==3.0.3
gunicorn==19.7.1
html5lib==0.999
idna==2.6
itsdangerous==0.24
Jinja2==2.10
language-selector==0.1
lxml==3.3.3
MarkupSafe==1.0
mccabe==0.6.1
mysqlclient==1.3.12
paramiko==2.4.0
pep8==1.7.1
pew==1.1.2
pipenv==9.0.3
pycodestyle==2.3.1
pycurl==7.19.3
pyflakes==1.6.0
pygobject==3.12.0
pyparsing==2.2.0
python-apt===0.9.3.5ubuntu2
rauth==0.7.3
requests==2.18.4
six==1.11.0
SQLAlchemy==1.1.15
ufw===0.34-rc-0ubuntu2
unattended-upgrades==0.1
urllib3==1.22
virtualenv==15.1.0
virtualenv-clone==0.2.6
Werkzeug==0.13
[requires]
python_version = "3.4.3"
